gym_system/desktop_software/requirements.txt
customtkinter>=5.2.0
pyodbc>=4.0.39
requests>=2.31.0
Pillow>=10.0.0
pyinstaller>=6.0.0
